requirements.txt
# Webserver/wsgi
gunicorn
meinheld
json-logging

# Flask & plugins
Flask
flask-cors
Flask-Migrate
flask-sqlalchemy

# General dependencies
requests
python-dotenv
pydantic
pyIsEmail
